Fix some aspects of pipe collision by removing extra additions of mDeltaP

+++ Utils.h
@@ -10,7 +10,7 @@ public:
     static bool IsOnLeft(sf::Vector2f a, sf::Vector2f b, sf::Vector2f c);
     static bool IsOnRight(sf::Vector2f a, sf::Vector2f b, sf::Vector2f c);
     static bool IsCollinear(sf::Vector2f a, sf::Vector2f b, sf::Vector2f c);
-    static int Area2( sf::Vector2f a,  sf::Vector2f b,  sf::Vector2f c);
+    static float Area2( sf::Vector2f a,  sf::Vector2f b,  sf::Vector2f c);
 };
 
 #endif  // SUPERMARIOBROS_UTILS_H
